updated json endpoints to not take body and extract params themselves, but the decorator will do that automatically and fixed a bug where a Child and Parent could share the same api key
